Move Edu definition into own file
