+++ ruma-federation-api/src/transactions/send_transaction_message.rs
@@ -1,3 +1,4 @@
 //! Endpoint to send live activity messages to another server.
 
+pub mod edu;
 pub mod v1;
